Move results pages for genre and related artists to pages with path params
It seems that Remix auth (or something) will periodically drop query
parameters when requesting the next URL. To get around this, those
result pages are moved to ones with path parameters which the old URLs
redirect to, such that paths stay alive.
