+++ app/components/Album/ReviewButtons.tsx
@@ -41,7 +41,7 @@ const ReviewButtons: React.FC<ReviewButtonProps> = ({ item }) => {
             <ButtonLinkGroup
               items={item.genres.slice(0, 3)}
               keyFunction={(genre, i) => `${genre}-${i}`}
-              toFunction={(genre) => `/genre?genre=${genre}`}
+              toFunction={(genre) => `/genre/${genre}`}
               childFunction={(genre) => genre}
               className={clsx('btn-xs')}
               wrapperClassName={clsx('mb-2')}
